#283 move all tests that require settle_bud to src/bud/bud_test
